New levels can be generated at will on pressing the return key
Unrelated to that, squares can effectively be "useless" if they only have a left arrow and are positioned at the beginning of the array, or only a right arrow and positioned at the end. I'm wondering if this is an interesting bit of obfuscation? Or will it be unnecessary "noise" in the player reading of a level?

If unnecessary, I should either remove that arrow or make it so that only applicable interactions are generated by the level creation algorithm