A solution/challenge is generated so that the play has a pattern of squares to match in a number of clicks
The solution is fully random for now (the number of clicks for the solution is also random within a range)

This is done because I have no idea of heuristics that indicate whether some aspects or a level's configuration (or its goal) make it more interesting than another, those will have to wait

? Should the solution/goal include the possibility of clicking the same square more than once?

Unrelated to the solution/goal thing but I should find  better way to indicate a self target than a "down" arrow

Also: I should settle on a single term for the solution/goal/challenge of a level. For documentation but also for my sanity. We'll see what sticks