Added a readme + process and design documentation
I figured it would be useful to have a record of the "instructions" for the game. Whether or not their included textually in the game, it'll be interesting to see how/if these evolve

(some additional thoughts on this: board games can afford to have explicit textual "external" instructions about how to play the game, but video game designer burden themselves with the task of teaching the players their game "from within" the game. Why is that? Can we be lazy and just handout instructions to our players instead of designing tutorials?)