Clicks counter + undo/redo/reset buttons
Adding the possibility of navigating the "history" of the puzzle being solved seemed important. However, after implementing these features and playing with them, I realized one thing: it's a bit unintuitive to navigate the history of a puzzle being solved if there's no "movement" within the puzzle

What I mean is that undoing or redoing a move feels weird when you can't actually see the "move" go in reverse. Now it just seems like another move is being made, instead of navigating the history of already-made moves

First time I'm actually seeing this hurdle in minimalism. I guess it's a consequence of making a game with no movement at all (think movement metaphor for interaction in Svanaes 2000)

Adding some animations or some indicators of "made moves" that can animate or backtrack could solve this