Added a feedforward/telegraphing feature to show the toggling predictions of clicking on a square
The intent of this is to ease gameplay/exploration with the level. I feel that knowing a bit more about what's going to happen as you're *about* to interact with a square will go a long way

This extra important in this kind of game because nothing is really telling you what's going to happen "naturally". Affordances are hidden, therefore feedback and feedforward and signifiers are doing all the heavy lifting

I'm pretty happy with the feel of it and the level of information it's giving. It's making interaction with the squares "richer", in that it supports anticipation, not just trial and error

I'm not happy with the look of it, though, but I'll revisit that some other time