Light level generation tuning + minor visual polishes
I wanted to get some builds ready and available for testing so I did some fine tuning like changing the range of clicks for objectives

Before, it could go as low as 2 on any level and that felt really underwhelming on longer arrays. I Changed it so it's always 1 fewer than the number of squares in a level. That decision was almost arbitrary, but the intention is that longer levels are consistently more difficult than smaller ones (I'm under the impression that more clicks == harder? I should probably test that hypothesis at some point)

Also changed made it so undoing and resetting is possible a when a level is completed. This opens up the possibility of players completing levels with different amounts of clicks. This way, we can reward them for specific solutions. Right now I decided that solving a level with the exact amount of clicks is harder/rewarded with a special animation (Just a change of color, really... Minimalism, you know. In a world of black and white and grey, green should be appreciated in a special way)

Finally, I decided to be thorough and explicit with the readme/how to play. So the game is playable by reading the complete set of instructions even though the in-game feedback and signifiers are not sufficient