Added some code to test and output all possible solutions toward a goal arrangement of squares
This brought to light a very important piece of information about generated levels and their solutions. The order in which the squares are clicked is not important! This is a huge eye-opening moment and I will document this and its implications further in the journal

This is part of a feature to allow/have players "play out" all the possible solutions of a level in a order to truly solve it