Displayed the number of clicks for every possible solution of a level on the UI
This indicates the broader goal of a level: playing out all possible solutions

It invites players to do so even though they might not have realized multiple solutions were possible

I might make this "feature" optional, like allowing players to go to the next level after completing only one solution, but rewarding players "more" (how???) if they complete all solutions before moving on to the next level

(I needed to decide on an order for the display, I went this highest to lowest as I feel this what players would intuitively do when going through multiple solutions, but when testing I immediately did the opposite... mmm)

((also should we limit possible solutions to only one per amount of clicks? e.g. only one with 3 clicks, even though multiple solutions with 3 clicks are possible? It feels a bit weird to have the same number twice on the list))

(((also should find a way to display/inform the player of the solutions that they previously succeeded at completing so they don't accidentally re-do the same one over and over (happened to me when testing), this is going to be tricky since the order of the clicks don't matter.... I'm thinking of "printing" the solutions as thin rectangles (indicating the squares involved in a solution, with no info on the order) under the array after they've been completed by the player)))