Implemented a working version of the multiple solution feature
In this "mode", the players have to find all solutions to a puzzle before moving on to the next

They do have the reset the puzzle manually in order to try out and complete new solutions. Maybe we should do this automatically through a quick animation

There is also a problem: when you have multiple solutions and, for example, they're of lengths 2 and 4, you can just do the one with length 2 + press any square twice and that'll "give" you the one of length 4 without actually finding the true length 4 solution. I can solve this by making the squares only able to be clicked once per level. This will limit me in considering new features, like the conditional toggles that I was thinking about. But it's a necessary change for now, we can always explore other solution that problem later