Made squares uninteractable on click
This is so squares are only able to be clicked once per level. This solves the issue of "cheating" multiple solutions by clicking a number of times on the same square

Not 100% happy with the visual feedback of it, a light "fading" of the toggle indicator. I can't really have a big overlay with a bar that crosses out a square or that obscures the color of a square because we still need to be able to see if a square is toggled on or off

We'll need to test if that indicated uninteractability adequately