Change the placeholder arrows for the UI history buttons
There was some confusion about the functionality of these button during the playtests. I assume the confusion mostly took place because I was still using the placeholder image files and they had little to no visual indication about what the button they were on did.

I changed to a more "universal" iconography for undo/redo/reset in the hopes of clearing that confusion. I also changed the color of the arrow when the button isn't interactable to make it more clear that it isn't (faded grey = does nothing)