Introduced the "wrap around" toggle features
Maybe a feature that will be a step in the level progression?

Someone suggested doing that during the feedback session last week. It was pretty quick to implement since it was just a matter of supporting the actual wrap around functionality + preview graphics with the existing "useless arrows" that I removed some time ago

This should add a touch of complexity on the otherwise very linear level solving experience