Started to work on settings menu, but shelving this for now
This is to enable players to choose different game modes, turn on or off features and tweak the level generation algorithm from the game itself

I imported the settings menu from another project, thinking I could quickly hook up some of the basic the options from this game onto it, but it turns out it's a bit more complicated than that so I'll be shelving this for later. I'm not sure if it's even necessary

The quick and dirty version of what I'm trying to do is just to build two different versions of the game. One in multiple solution mode, the other in single solution mode