Added a very rudimentary progression scheme + questions document
So the progression only goes up linearly by added a square to the array every time players complete two levels. It will definitely be fleshed out more in the future but right now we have a semblance of progression + some UI feedback for how many levels cleared

There are debug controls to help with testing the progression mechanisms

I also added a "questions" document to compile the questions I ask myself about the game but that don't immediately translate into a todo entry or a reflection for the journal. I went back and added the questions I had in the past as well to show the evolution of the questions that pop up during development

BUT more importantly, the level generation algorithm is busted.... I don't know when this happened but the algorithm is spitting out impossible solutions... so I'm going to have to focus on this for a while since it's pretty important!