Fixed the broken level generation bug
Solutions with the same amount of steps were compiled and suggested to the players when they shouldn't have and some of the solutions suggested weren't even possible to get to!

I fixed this, but now that multiple solutions are correctly enforced in multi-solution mode, I am now realizing that some level setups are just impossible to generate multiple solutions for. This is a pretty big design problem that calls for a longer reflection so I'll be writing in the journal about that