More easily editable progression with more features attached to it + bug fix
So the mechanisms for adding features gradually to the game through progression is in. I'm slowly introducing players to "more complex" target arrows now. It's hard to know if the base values that I chose are good enough so I'll need to test that

Also to not that having more restrictions on which arrows to choose from in the generation of a level does hinder the level generation algorithm and puts a high dependence on the "pull from file if can't generate a valid level" system

This also might mean that we have to accept having single solution level in multi-solution mode at first. I feel that if it's just the first levels and that single-solution levels don't show up after multi-solutions are introduced, it doesn't feel too inconsistent and is acceptable. This would even feel like accessing multi-solution level is part of the progression which is a nice accident

But we'll have to test for consistency