Bug fixes + journaling and todo editing + updated builds for testing
Fixed the issue of duplicating/non-destroying solution clicks boxes and the non-disappearing next level button

Edited some todo entries to record ideas I had about how to implement some features. I'm not sure if it's the right place to put those reflections/thoughts but since they're directly related to work items I intend on doing soon-ish, it made sense to write them there

I'm thinking: if something is partly implemented and I have thoughts about a further expansion/modification of that feature, I'll write these thoughts as in-code comments near the part(s) of the code that relate to that feature

If something isn't implemented and I have thoughts about a future/possible/potential feature, I'll write these thoughts in the todo entries about that feature

I also built updated versions of the two modes so they can be shared online and played/tested. The main update is the progression through targeting schemes + wrap around toggles and the "forced" single/multi solution that wasn't as robust and consistent before. Cascading toggles are also in but they're not meant as the focus of the next playtest

Oh and I also receive some spontaneous bits of feedback about the possibility of brute forcing levels, a suggestion to be able to "go back" and revisit prior levels and some preference for the single solution with its "hard edges"

I'd love to address the brute force issue but, to be honest, I'm not exactly sure it's possible to fix... maybe by throwing longer arrays at the players earlier? Maybe only smaller levels are easy to brute force? This is a little bit of a dead end for me as of right now... I'll have to revisit that issue later

The "revisit previous levels" feature sounds nice to me and shouldn't be too hard to implement, but I'm wondering about its value. Maybe it would help and support players in their learning of the game? Apart from that, I'm not sure of what it would bring to the experience

Finally, the mention of a satisfying red shaking response in single solution mode is interesting. I might be able to bring something similar to multi-solution mode? I'm not sure if it's compatible with this mode yet but I can try and few things to try and get the "best of both worlds", since I feel like I have a slight preference for multi-solution mode but, indeed, the feedback from busting the clicks counter in single-solution is pretty satisfying