Wrote a simple function to sort the levels by square lengths in the prevalidated levels files
The point of this is to maximize the number of prevalidated levels we can stack up while testing/developing the game without having to order them manually

There still is a button we have to press. I might hook up the function to an automatic process at some point but since we also have to manually transfer the files from my local PC data to the project's resource folder in order for the levels to be built into the game, this will do for now. (That whole process can be automated at some point too, but we'll see if it's necessary later)

Right now I'll wait to have more levels with cascading toggles that are prevalidated before transfering the files to be built