Preliminary setup for including cascading toggles in progression + typing up feedback entries
Received feedback from playtests and typed it up to address it. There isn't a clear "best mode" between multi and single solution. I hoped the playtests would help me decide on which mode to keep in the game but I received very good pros and cons for both approaches. The reception was slightly more negative for the multi-solution mode. Right now I'm wondering if there's anything I could "salvage" from the multi-solution mode to bring into single solution mode as a "best of both worlds" approach.

The other solution might be to progress from one to the other. We could test this soon as my next immediate task is to rework/polish the progression system