Continued the setup for the progression system overhaul + removed the possibility for squares that aren't targeted by others to be cascading
The rationale for the progression system rework is the following: right now the level generation algorithm works in a way that levels are generated with random features that the progression system *allows* to be generated. This means that there is a possibility for a very late level, which allows for very advanced features, to have none of these features because of the random aspect. I want to change it to make it so features can be expected and ensured to be present on specific levels. More on that later

In the meantime, I wanted to exclude superfluous features like squares indicated as cascading that can't cascade because they can't be target by other squares. This is important because having a cascading indicator on something that can't cascade feels like a false signifier (a false affordance?) and the promise of well communicated, discoverable and usable interactive design would be somewhat broken