Continued the setup for the progression system overhaul
The change to the progression system is going to require a major change to the level generation algorithm. Changing such a convoluted system is daunting so I started the process by reviewing the existing system to get a solid grasp on how it works so I can put my hands in the wires and pull what's necessary without breaking everything. I wrote a thorough block of pseudocode that goes through everything the algorithm does to generate a level, its squares and their features, its goal and its multiple possible solutions. After doing this, I feel that I have the necessary understanding of the system to actually change it, now