Last(?) tweaks to the progression system
This is the last "feature" I'll work on. Not exactly a new feature but I wanted to introduce the very basic mechanisms of the game through the progression system

Now players will start with a one-self-targeting-square level to learn the main mechanic of the game: clicking a square to toggle it on/off. The second level will be a 2-squares level with an side-arrow and a self-targeting square

My intention with these two levels is to make sure players understand how to interact with the game as its core (and how these interaction lead to the objective of the game) before actually trying solve challenges

I also tweaked a few things in the "save level to file" system so that we don't have duplicate levels in the files and that there can't be level with all self-targeting squares (except the first level) because even though the first levels are meant to be simple, that's just plain boring