Tweaked progression to introduce cascading toggles on less complex targeting schemes first + experimented with different target prediction graphics (including the implementation of a version that supports/includes cascading toggle targets
My intentions here were again to smooth the difficulty curve of the game so that when the hardest feature of the game is introduced, it's not coupled with some other hard features. It was a lot easier to do than anticipated, I managed to build a progression "editing" system that allows for a lot of variety and control over which features are included when and it supported the changes I wanted to do here. Very satisfying

This should have been two different commits, but I also played around with the prediction graphic/feedback. I had received a comment at some point, saying that the prediction rectangle were revealing too much about the game. I made them a bit more cryptic, and also managed to display them in a way that supports the cascading feature (because they are now smaller and toggle/color agnostic)

I am excited to test these (new "final" progression + new prediction graphics) with new players