Added more obvious feedback for the clicks counter and the level cleared counter
A while ago I had received the feedback that the clicks counter was a bit too easy to ignore and gloss over. I made it bigger and added a little punch animation to make sure the play notices it when it changes, so its function is more obvious

This should help players learn the mechanisms of the game quicker

I also added a similar feedback to the levels cleared counter, to again help the players understand the function of the numbers they see across the screen. A few commits ago I removed the trophy graphic that was helping to illustrate that counter's function because it broke the minimalist intention of not connecting any game rule and mechanism to real world analogs. I'm hoping that the animation connects the number to the in-game moment of clearing a level