Added the feedback I received from the latest playtest, along with todo entries that reflect what to do to address the feedback
I received a lot of positive feedback on the overall experience, progression and difficulty curve. Also about the general appeal and enjoyment of the game

The notes I took were mostly about the "negative" things. These were mostly small things that can be addressed with visual tweaks or adjustments on the progression, not necessarily big changes in the code. I'm glad