Made the second level only one click so it's not pre-solved
This is to address the point of feedback I received about it feeling weird that the second level was (and players felt like it was possible that others could also be) pre-solved.

This is good for consistency. I had originally thought that having a single exception wouldn't bother players but apparently it did

Now all of the levels follow the exact same rules, no exception