Made the level completion animation play every time a level is re-solved, with a bit of a difference
Players felt doubtful that they really solved a level if they didn't see the animation play every time they entered the correct solution. I previously deliberately stopped the animation from playing repeatedly to communicate that the level was already solved, but now I realized it created confusion.

I reinstalled the animation to be played repeatedly but changed it a bit. I made the graphic thinner and less colorful every time it's animated after the first time. That way, players can still be confirmed that the right solution is entered, but they can also see that the animation is different, indicating that the level is indeed already solved