Hid the redo/reset/undo buttons until a square is pressed for the first time
The intention here is to isolate the core interaction of the game as much as possible as the first thing the player should do.

Some time ago I saw some players hesitate and try and click all of the buttons that appeared on screen, including the UI buttons. To minimize that, I'm trying to hide them before the first square is clicked. I hope that this communicates to the player that those buttons are secondary to the middle buttons and that they are not integral to the gameplay