Made it a bit more complicated to use the debug functions
I added an extra key to press in order to use the debug functions. A player stumbled upon those functions before and it basically broke the game/the experience for them so I made it so you need a to keep the D key pressed to use them. The intention is to make them impossible to trigger by accident

I also fixed a few bugs I had with the debug function and the resetting of early levels