Added checks and crosses (+ changed the general position of the solution rectangles + added feedback on click/toggle) to help players read levels and the game's objective
The main piece of feedback I got from that last playtest was that the objective is not obvious when playing for the first time. After a discussion with the testers, I realised that this was mostly due to the solution rectangles being a bit to far away from the array of interactive square and that there wasn't enough feedback to connect what players were doing with the squares to the solution rectangles and therefore the objective of the level/game

By adding the checks and the crosses, my intention is to make it really obvious that these icon represent completion, and therefore are connecting what they're close to to the objective of the game. By moving the solution rectangles up and squeezing the checks and crosses between the two connected array, I'm hoping that the connection will be obvious

If it's still not obvious enough, I doubled down by adding bits of extra feedback when clicking and toggling squares. Squares now "bump" when toggled, along with their respective solution rectangle and even with the associated check or cross

With this, every connected object of the game animates at the same time, connecting the action the player has done to the game's reaction in a much more visually informative way

(I ran a small and simple survey with a few friends to figure out which layout of solution rectangles looked best and was more intuitive: above the squares or below. Above got 1 vote, below got 8)