Added an animation + a delay for toggling
The latest feedback I got from having someone new try the game for the first time was that it's still not super clear what's happening. Even with the new animations, checks and crosses to emphasize mappings between different elements, it's still not great

I went back to the feedback I had received from past playtest sessions and revisited an old idea: to add a "nudge" animation to the arrow(s) of a clicked square to make the relationship between the squares and the "targeted toggles" more obvious

After having done that, I basically had to also implement a way to animation/separate/delay the cascading toggles, since it's basically an extension of the same feature

I also decided to remove the target prediction dots. Someone told me during last playtests that, even though they were meant to help in understanding the mechanics and relationship between the squares, they added noise to the screen and made things muddier. This struck a chord, since by making minimalist games I'm of course trying to minimize the noise and visual clutter. I believe that by adding the nudge animations and delays, these dots wouldn't be necessary anyway