More tweaks on animations
Many minor tweaks on speed and delays here and there but specifically, I added a delay on the level complete animation to try and capture the attention of the player "sequentially" so that they look at things happening one by one instead of all at once

Oh and I changed the layout of the solution rectangles, which are now squares

This is to follow through with the idea that I wasn't being minimalistic enough. Checks, crosses and bumping rectangles were cluttering the visual space. Now with the delays and the animation MINUS the visual noise, I'm hoping the connection between the middle interactive squares and the bottom object squares is even clearer