Fixed a bug and a semi-bug + added feedback from last play session
Addressed a couple of issues that came up during the last feedback session.

One was a bug I had before and thought I fixed, I fixed it for good

The other was that multi-solution levels sneaking their way into the single solution mode (the actual game) felt buggy, so I figured that, since I can't always force single solution levels, I'll display the objective click counter for a level to show the solution with the least amount of clicks. That way, there might be more possible solutions to a level than the one the player achieved, but they won't know it. It's a cheap fix, but it should work

The other issue is that some levels are impossible. I investigated and, indeed, some levels are impossible to solve even with the provided solution... this is bothersome. They don't happen often but it's a problem for sure... I'll need to dig deeper for this one...

The other bits of feedback I received was that it was a good call to remove the prediction dots, to remove the checks and crosses, and to make the objective array squares instead of rectangles