Fixed the major bug that allowed the generation of impossible levels
There was an issue with the way levels were generated after I added the animations and delays. I took a wild swing at troubleshooting it and managed to isolate what the issue was (I wasn't turning off cascading for the right squares when fake-clicking them to generate and display a solution)

I'm pretty happy with this. This was the last problem/major issue that I wanted to tackling before calling the game done

...

Of course I still have some tiny features I want to implement still... just some minor visual (and audio) things...

I was thinking a cute color change when beating levels (or every couple of levels, like when reaching a new mechanic), or a counter to indicate how many levels are part of the "official" progression. Oh and adding some audio feedback

We'll see if that happens or not! If not, I'm done