Attempt at playing with colors + indicating the "total" amount of levels
I set up a way to swap in/out the color palette of the entire game. My intention is to use (monochromatic) color changes to reward the player for beating a certain amount of levels/indicate a shift in progression (when new mechanics are introduced, for example)

I've not yet decided on actual color schemes to implement but the system is there. This is the kind of decision that I don't like to make (i.e. deciding which colors to use), hence the original shades of grey. Making "concrete" decisions is hard, and I don't like doing that. I'd rather make abstract, systemic, algorithmic, pragmatic decisions that, in turn, affect the game's design (both aesthetically and gameplay-wise)

Still, the value of using color in the game is non-negligible, so I'm trying to outsource a little bit and find and get people motivated to pitch color schemes to use in the game. Another way to do this would be to find "pretty colors schemes" algorithmically, but I currently don't have a specific intuition how to approach this

This should have been two commits:

The second thing I did here is experiment with ways to communicate the "total" amount of levels in the game. Players ask me this all the time. Or similar questions like: "how many levels are there?", "How long is this?", etc.

It's not obvious at first that levels are procedurally generated so players don't know there's literally no end to this game. However, after 40 levels, no new mechanic is introduced and no change in the rules of configuration of the levels  is making the game harder, so I consider that, by beating 40 levels, you've functionally seen all that the game has to provide. It's still very much possible to continue playing, but you've pretty much "beaten" the game after 40 levels.

So I've tried adding a "/40" label to indicate that. Although I'm not a fan of integrating more visual information on screen, there should be a way to make this as minimal as possible

I've tried having that label fade and disappear after "informing" the player. I've tried making is part of the "levels cleared counter box", I've also tried having a non-numerical indicator (little squares filling up a big square). No one solution is sticking out as the best one yet. I've sent out screenshots and clips of multiple versions of this to father feedback about the preferred solution. To be continued...